pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "django-oscar-announcements"
version = "0.1.0"
description = "Pinax Announcements integration for Django Oscar dashboards"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "django-oscar",
    "pinax-announcements>=4.0",
]

classifiers = [
    # https://pypi.org/classifiers/
    "Development Status :: 4 - Beta",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

[project.urls]
Documentation = "https://github.com/niccokunzmann/django-oscar-announcements"
Repository = "https://github.com/niccokunzmann/django-oscar-announcements"
Changelog = "https://github.com/niccokunzmann/django-oscar-announcements/blob/main/CHANGES.md"


[tool.setuptools.packages.find]
where = ["."]
include = ["oscar_announcements*"]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-django",
    "pytest-mock",
    "pre-commit",
]

[tool.setuptools.package-data]
oscar_announcements = [
    "templates/**/*.html",
    "static/**/*.js",
    "static/**/*.css",
]
